Created the starter files for this project
The sidebar is finished. Perhaps I will add more styling later and I haven't added the logo. The topnav is also semi-finished. I still need to figure out how to remove the margin-right on the icon class on the login button. If I remove it from the css, then it will also be removed from the buttons on the sidebar.